Fix feature errors
