+++ keriox_core/src/event_message/mod.rs
@@ -1,7 +1,6 @@
 pub mod cesr_adapter;
 pub mod dummy_event;
 pub mod event_msg_builder;
-pub mod exchange;
 pub mod key_event_message;
 pub mod serialization_info;
 pub mod serializer;
