+++ keriox_core/src/event_message/signed_event_message.rs
@@ -1,7 +1,7 @@
 use serde::{ser::SerializeStruct, Deserialize, Serialize};
 
 use super::{
-    cesr_adapter::ParsedEvent, exchange::SignedExchange, key_event_message::KeyEvent,
+    cesr_adapter::ParsedEvent, key_event_message::KeyEvent,
     serializer::to_string, signature::Nontransferable, EventMessage,
 };
 #[cfg(feature = "query")]
@@ -14,8 +14,10 @@ use crate::{
     },
     event_parsing::{group::Group, ParsedData},
     prefix::{AttachedSignaturePrefix, IdentifierPrefix},
-    state::{EventSemantics, IdentifierState},
-};
+    state::{EventSemantics, IdentifierState}}; 
+
+#[cfg(feature = "mailbox")]
+use crate::mailbox::exchange::SignedExchange;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Message {
@@ -34,10 +36,11 @@ pub enum Notice {
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Op {
+    #[cfg(feature = "mailbox")]
     Exchange(SignedExchange),
-    #[cfg(feature = "query")]
-    Reply(SignedReply),
     #[cfg(any(feature = "query", feature = "oobi"))]
+    Reply(SignedReply),
+    #[cfg(feature = "query")]
     Query(SignedQuery),
 }
 
@@ -67,6 +70,7 @@ impl From<Op> for ParsedEvent {
             Op::Reply(ksn) => ParsedData::from(ksn),
             #[cfg(feature = "query")]
             Op::Query(qry) => ParsedData::from(qry),
+            #[cfg(feature = "mailbox")]
             Op::Exchange(exn) => ParsedData::from(exn),
         }
     }
@@ -104,8 +108,10 @@ impl Op {
             Op::Reply(reply) => reply.reply.get_prefix(),
             #[cfg(feature = "query")]
             Op::Query(qry) => qry.query.get_prefix(),
+            #[cfg(feature = "mailbox")]
             // returns exchange message receipient id
             Op::Exchange(exn) => exn.exchange_message.event.content.data.get_prefix(),
+            _ => todo!(),
         }
     }
 }
